+++ src/Product/Data/ProductCrosssellRelation.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Lsv\Datapump\Product\Data;
+
+class ProductCrosssellRelation extends ProductRelation
+{
+    public function getKey(): string
+    {
+        return 'cs_skus';
+    }
+}