+++ src/Product/Data/ProductUpsellRelation.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Lsv\Datapump\Product\Data;
+
+class ProductUpsellRelation extends ProductRelation
+{
+    public function getKey(): string
+    {
+        return 'us_skus';
+    }
+}